refactor: 重构导航菜单、右键菜单的集成方式
