+++ src/components/workflow/menu/GraphContextMenuIntergration.ts
@@ -1,0 +1,11 @@
+
+import { MenuIntegerationArgs } from "./Menus";
+import { useGraphNavMenu } from "./GraphNavMenuIntergration";
+import { useBlankContextMenu } from "./BlankContextMenuIntergration";
+
+
+export function useGraphContextMenu(params: MenuIntegerationArgs) {
+    useGraphNavMenu(params);
+    useBlankContextMenu(params);
+}
+
