--- src/components/workflow/menu/MenuItem.ts
@@ -1,11 +1,0 @@
-
-
-/** 
- * 菜单项
- */
-export default interface MenuItem {
-    menuId: string,
-    menuIcon: string,
-    menuName: string,
-    menuCallback: (id: string) => void
-}
